Correct createHistoryName argument name
Previously the example and the table documenting the arguments used
conflicing names. This commit attempts to correct that and document what
is actually going on.

Related to https://github.com/ssbc/scuttlebutt-protocol-guide/pull/38. I
confirmed that Patchwork uses `seq` and not `sequence` when producing
messages however it seems that the code usually accepts both of those
arguments.

Given that `sequence` seems to be preferred we should standarise on it
in the future and document this behaviour.